Move OrchestratorClient and Watch facilities to a distinct module
