+++ ts/libs/core/src/index.ts
@@ -6,6 +6,7 @@ import 'reflect-metadata';
 
 export * from './lib/util/public';
 export * from './lib/model';
+export * from './lib/orchestrator/public';
 export * from './lib/raw-metrics-query/public';
 export * from './lib/composed-metrics/public';
 export * from './lib/elasticity/public';
@@ -15,6 +16,7 @@ export * from './lib/runtime/public';
 
 // We want to make these implementations public, but they are not included in the respective `xyz/public/index.ts` files,
 // because they would create circular dependencies.
+export * from './lib/orchestrator/public/impl';
 export * from './lib/raw-metrics-query/public/time-series/execution/impl';
 export * from './lib/composed-metrics/public/common/impl';
 export * from './lib/composed-metrics/public/control/impl';
