+++ ts/libs/core/src/lib/elasticity/public/control/impl/base/vertical-elasticity-strategy-controller.base.ts
@@ -10,7 +10,8 @@ import {
     SloTarget,
     VerticalElasticityStrategyConfig,
 } from '../../../../../model';
-import { OrchestratorClient, PolarisRuntime } from '../../../../../runtime';
+import { OrchestratorClient } from '../../../../../orchestrator';
+import { PolarisRuntime } from '../../../../../runtime';
 import { Logger } from '../../../../../util';
 import { ElasticityStrategyExecutionError, StabilizationWindowTracker } from '../../../common';
 import { DefaultStabilizationWindowTracker } from '../default-stabilization-window-tracker';
