+++ ts/libs/core/src/lib/orchestrator/public/impl/index.ts
@@ -1,0 +1,1 @@
+export * from './default-watch-manager';
