+++ ts/libs/core/src/lib/orchestrator/public/index.ts
@@ -1,4 +1,5 @@
-export * from './errors';
+export * from './error';
 export * from './object-kind-watcher';
+export * from './orchestrator-client';
 export * from './watch-events-handler';
 export * from './watch-manager';
