+++ ts/libs/core/src/lib/runtime/public/impl/polaris-runtime.base.ts
@@ -3,18 +3,17 @@ import { DefaultComposedMetricsManager } from '../../../composed-metrics/public/
 import { ElasticityStrategyManager, ElasticityStrategyService } from '../../../elasticity';
 import { DefaultElasticityStrategyManager } from '../../../elasticity/public/control/impl/default-elasticity-strategy-manager';
 import { DefaultElasticityStrategyService } from '../../../elasticity/public/service/impl/default-elasticity-strategy.service';
+import { ObjectKindWatcher, OrchestratorClient, WatchManager } from '../../../orchestrator';
+import { DefaultWatchManager } from '../../../orchestrator/public/impl';
 import { SloControlLoop, SloEvaluator } from '../../../slo';
 import { DefaultSloControlLoop } from '../../../slo/public/control/impl';
 import { PropertyTransformer } from '../../../transformation/internal/property-transformer';
 import { PolarisTransformationServiceManager } from '../../../transformation/public/common';
 import { DefaultPolarisTransformationService } from '../../../transformation/public/impl';
 import { MetricsSourcesManager } from '../metrics-source';
-import { OrchestratorClient } from '../orchestrator-client';
 import { PolarisRuntime } from '../polaris-runtime';
-import { ObjectKindWatcher, WatchManager } from '../watch';
 import { DefaultMetricsSourcesManager } from './default-metrics-sources-manager';
 import { DefaultMicrocontrollerFactory } from './default-microcontroller-factory';
-import { DefaultWatchManager } from './default-watch-manager';
 
 /**
  * `PolarisRuntimeBase` can be used as a superclass for orchestrator-specific {@link PolarisRuntime} implementations.
