+++ ts/libs/core/src/lib/runtime/public/index.ts
@@ -1,4 +1,3 @@
 export * from './metrics-source';
-export * from './orchestrator-client';
+export * from './microcontroller-factory';
 export * from './polaris-runtime';
-export * from './watch';
