--- ts/libs/core/src/lib/runtime/public/orchestrator-client/index.ts
@@ -1,2 +1,0 @@
-export * from './error';
-export * from './orchestrator-client';
